Rewritting of the local HTML parser
